requirements.txt
asgiref==3.8.1
Django==5.2.1
django-role-permissions==3.2.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
whitenoise==6.7.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
